requirements.txt
streamlit
radon
gitpython
networkx
matplotlib
pydot
graphviz  
pyvis
tree_select